+++ NgxExport/Tools/PCRE.hs
@@ -14,7 +14,7 @@
 -- Portability :  non-portable (requires Template Haskell)
 --
 -- PCRE matching and substitution from the more extra tools collection
--- for <http://github.com/lyokha/nginx-haskell-module nginx-haskell-module>.
+-- for <https://github.com/lyokha/nginx-haskell-module nginx-haskell-module>.
 --
 -----------------------------------------------------------------------------